services/product_service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp-proto-http==1.29.0
prometheus-fastapi-instrumentator==6.1.0
